equipments controller
equipment controller to handle crud operations.
